Code cleanup, renamed files and moved some ranking helper functions
