+++ ranker/ranker.py
@@ -1,16 +1,18 @@
 from flask import Flask, jsonify, request
-
-import requests
 import sys
-import time
 sys.path.append('..')
 sys.path.append('../crawler')
-import helpers
-log = helpers.log
+
+import requests
+import time
+
+from helpers import *
 
 from mongoengine import *
 from mongoConfig import *
-from cosineSimilarity import *
+
+from rankDocuments import rank
+from stemQuery import stemQuery
 import rankerDBConfig
 from loadInvertedIndexToMemory import loadInvertedIndexToMemory
 
@@ -29,8 +31,8 @@ def index():
 def rankQuery(query):
   log('Ranker', 'Received query: '+query)
   queryTerms = stemQuery(query)
-  sortedDocUrls=calculateRankings(queryTerms, inMemoryTFIDF, crawlerReverseMap, termReverseMap, pageRanks, authority)
-  return helpers.sendPacket(1, 'Successfully retrieved query', {'sortedDocUrls':sortedDocUrls})
+  sortedDocUrls=rank(queryTerms, inMemoryTFIDF, crawlerReverseMap, termReverseMap, pageRanks, authority)
+  return sendPacket(1, 'Successfully retrieved query', {'sortedDocUrls':sortedDocUrls})
 
 @app.route('/fetchDocuments', methods=['POST'])
 def fetchDocuments():
@@ -42,7 +44,7 @@ def fetchDocuments():
     documents.append(Crawler.objects.get(url=url).to_json())
 
   log('ranker', 'Finished fetching documents in '+str(time.time() - startTime) + ' seconds')
-  return helpers.sendPacket(1, 'Successfully retrieved documents', {'documents':documents})
+  return sendPacket(1, 'Successfully retrieved documents', {'documents':documents})
 
 if __name__ == "__main__":
   log('info', "Ranker is listening on port "+str(port) +", " + str(app.config['ENV']) + " environment.")
